Rename `3DS` to `AuthorizingPayment`
